requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
